[SPARK-3724][MLLIB] RandomForest: More options for feature subset size.
Move tests from mllib to ml. Replace extractors with regex.
